Set submission to "validating" once a worker fetches it.
This solves the problem where, in the context of multiple workers,
multiple workers may receive the same task.

Note that this creates a potential issue where, when a worker dies after
fetching the task and before submitting the results, the task will be
"orphaned" and requires manual rejudging. Solving this issue may
require, say, a separate field in the database recording the last time a
task is sent and retrying those that seem to be stale.